+++ crates/rtlola-io-plugins/tests/byte_plugin/structs/input_map.rs
@@ -1,14 +1,11 @@
-use std::io::ErrorKind;
 use std::time::Duration;
 
-use byteorder::ByteOrder;
 use rtlola_interpreter::config::OfflineMode;
 use rtlola_interpreter::input::{EventFactoryError, InputMap, MappedFactory};
 use rtlola_interpreter::monitor::Incremental;
 use rtlola_interpreter::time::{AbsoluteFloat, TimeRepresentation};
 use rtlola_interpreter::{ConfigBuilder, Monitor, Value};
 use rtlola_io_plugins::byte_plugin::time_converter::TimeConverter;
-use rtlola_io_plugins::byte_plugin::{ByteParsingError, FromBytes};
 use serde::{Deserialize, Serialize};
 
 use super::SPEC;
@@ -62,25 +59,25 @@ impl InputMap for TestInput {
     }
 }
 
-impl<B: ByteOrder> FromBytes<B> for TestInput {
-    type Error = <Self as InputMap>::Error;
+// impl<B: ByteOrder> FromBytes<B> for TestInput {
+//     type Error = <Self as InputMap>::Error;
 
-    fn from_bytes(data: &[u8]) -> Result<(Self, usize), ByteParsingError<<Self as InputMap>::Error>>
-    where
-        Self: Sized,
-    {
-        let res: TestInput = bincode::deserialize(&data).map_err(|e| {
-            if matches!(ErrorKind::UnexpectedEof, _e) {
-                ByteParsingError::Incomplete
-            } else {
-                ByteParsingError::Inner(EventFactoryError::Other(e))
-            }
-        })?;
-        let size =
-            bincode::serialized_size(&res).map_err(|e| ByteParsingError::Inner(EventFactoryError::Other(e)))? as usize;
-        Ok((res, size))
-    }
-}
+//     fn from_bytes(data: &[u8]) -> Result<(Self, usize), ByteParsingError<<Self as InputMap>::Error>>
+//     where
+//         Self: Sized,
+//     {
+//         let res: TestInput = bincode::deserialize(&data).map_err(|e| {
+//             if matches!(ErrorKind::UnexpectedEof, _e) {
+//                 ByteParsingError::Incomplete
+//             } else {
+//                 ByteParsingError::Inner(EventFactoryError::Other(e))
+//             }
+//         })?;
+//         let size =
+//             bincode::serialized_size(&res).map_err(|e| ByteParsingError::Inner(EventFactoryError::Other(e)))? as usize;
+//         Ok((res, size))
+//     }
+// }
 
 impl TimeConverter<AbsoluteFloat> for TestInput {
     fn convert_time(&self) -> Result<<AbsoluteFloat as TimeRepresentation>::InnerTime, <Self as InputMap>::Error> {